flq_simulation/requirements.txt
# FLQ-Fed 联邦学习量化框架依赖

# 深度学习框架
torch>=2.0.0
ultralytics>=8.0.0

# 联邦学习通信
fastapi>=0.95.0
uvicorn>=0.20.0
requests>=2.28.0

# 科学计算
numpy>=1.24.0
pyyaml>=6.0

# 可视化（可选）
matplotlib>=3.7.0
pandas>=2.0.0

# 测试框架（可选）
pytest>=7.0.0
pytest-cov>=4.0.0

# 开发工具（可选）
black>=23.0.0
flake8>=6.0.0

